+++ test/integration/fixture/repositories/answerRepository.js
@@ -1,9 +1,14 @@
 const THE_UNIVERSAL_ANSWER = 42
 
-module.exports = function({ timeout }) {
+module.exports = function({ timeout, conn }) {
   /* istanbul ignore next */
   if (!timeout) {
     throw new Error('No timeout specified')
+  }
+
+  /* istanbul ignore next */
+  if (!conn) {
+    throw new Error('No conn specified')
   }
 
   const getAnswerFor = function(question) {